Revert "Fix test package names & imports"
This reverts commit 99a8fe8fc2963c673dea915be543426e0a5a76b1.
