+++ data/src/test/java/org/oppia/android/data/gae/NetworkInterceptorTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.gae
+package org.oppia.android.data.gae
 
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
@@ -11,6 +11,8 @@ import dagger.Provides
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.oppia.android.data.backends.gae.NetworkInterceptor
+import org.oppia.android.data.backends.gae.NetworkModule
 import org.oppia.android.testing.ApiUtils
 import org.oppia.android.testing.network.MockTopicService
 import org.robolectric.annotation.LooperMode
