+++ data/src/test/java/org/oppia/android/data/gae/api/QuestionPlayerServiceTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.gae.api
+package org.oppia.android.data.gae.api
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
@@ -6,9 +6,10 @@ import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.oppia.android.testing.network.MockQuestionPlayerService
 import org.oppia.android.data.backends.gae.NetworkInterceptor
 import org.oppia.android.data.backends.gae.NetworkSettings
-import org.oppia.android.testing.network.MockSubtopicService
+import org.oppia.android.data.backends.gae.api.QuestionPlayerService
 import org.robolectric.annotation.LooperMode
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -16,11 +17,11 @@ import retrofit2.mock.MockRetrofit
 import retrofit2.mock.NetworkBehavior
 
 /**
- * Test for [SubtopicService] retrofit instance using [MockSubtopicService]
+ * Test for [QuestionPlayerService] retrofit instance using [MockQuestionPlayerService]
  */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
-class SubtopicServiceTest {
+class QuestionPlayerServiceTest {
   private lateinit var mockRetrofit: MockRetrofit
   private lateinit var retrofit: Retrofit
 
@@ -42,14 +43,19 @@ class SubtopicServiceTest {
   }
 
   @Test
-  fun testSubtopicService_usingFakeJson_deserializationSuccessful() {
-    val delegate = mockRetrofit.create(SubtopicService::class.java)
-    val mockSubtopicService = MockSubtopicService(delegate)
+  fun testQuestionPlayerService_usingFakeJson_deserializationSuccessful() {
+    val delegate = mockRetrofit.create(QuestionPlayerService::class.java)
+    val mockQuestionPlayerService = MockQuestionPlayerService(delegate)
 
-    val subtopic = mockSubtopicService.getSubtopic("Subtopic 1", "randomId")
-    val subtopicResponse = subtopic.execute()
+    val skillIdList = ArrayList<String>()
+    skillIdList.add("1")
+    skillIdList.add("2")
+    skillIdList.add("3")
+    val skillIds = skillIdList.joinToString(separator = ", ")
+    val questionPlayer = mockQuestionPlayerService.getQuestionPlayerBySkillIds(skillIds, 10)
+    val questionPlayerResponse = questionPlayer.execute()
 
-    assertThat(subtopicResponse.isSuccessful).isTrue()
-    assertThat(subtopicResponse.body()!!.subtopicTitle).isEqualTo("Subtopic 1")
+    assertThat(questionPlayerResponse.isSuccessful).isTrue()
+    assertThat(questionPlayerResponse.body()!!.questions!!.size).isEqualTo(1)
   }
 }
