+++ data/src/test/java/org/oppia/android/data/gae/api/StoryServiceTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.gae.api
+package org.oppia.android.data.gae.api
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
@@ -6,9 +6,10 @@ import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.oppia.android.testing.network.MockStoryService
 import org.oppia.android.data.backends.gae.NetworkInterceptor
 import org.oppia.android.data.backends.gae.NetworkSettings
-import org.oppia.android.testing.network.MockExplorationService
+import org.oppia.android.data.backends.gae.api.StoryService
 import org.robolectric.annotation.LooperMode
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -16,11 +17,11 @@ import retrofit2.mock.MockRetrofit
 import retrofit2.mock.NetworkBehavior
 
 /**
- * Test for [ExplorationService] retrofit instance using [MockExplorationService]
+ * Test for [StoryService] retrofit instance using [MockStoryService]
  */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
-class ExplorationServiceTest {
+class StoryServiceTest {
   private lateinit var mockRetrofit: MockRetrofit
   private lateinit var retrofit: Retrofit
 
@@ -42,14 +43,14 @@ class ExplorationServiceTest {
   }
 
   @Test
-  fun testExplorationService_usingFakeJson_deserializationSuccessful() {
-    val delegate = mockRetrofit.create(ExplorationService::class.java)
-    val mockExplorationService = MockExplorationService(delegate)
+  fun testStoryService_usingFakeJson_deserializationSuccessful() {
+    val delegate = mockRetrofit.create(StoryService::class.java)
+    val mockStoryService = MockStoryService(delegate)
 
-    val explorationContainer = mockExplorationService.getExplorationById("4")
-    val explorationContainerResponse = explorationContainer.execute()
+    val story = mockStoryService.getStory("1", "randomUserId", "rt4914")
+    val storyResponse = story.execute()
 
-    assertThat(explorationContainerResponse.isSuccessful).isTrue()
-    assertThat(explorationContainerResponse.body()!!.explorationId).isEqualTo("4")
+    assertThat(storyResponse.isSuccessful).isTrue()
+    assertThat(storyResponse.body()!!.storyTitle).isEqualTo("Story 1")
   }
 }
