+++ data/src/test/java/org/oppia/android/data/gae/api/SubtopicServiceTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.gae.api
+package org.oppia.android.data.gae.api
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
@@ -6,9 +6,10 @@ import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.oppia.android.testing.network.MockSubtopicService
 import org.oppia.android.data.backends.gae.NetworkInterceptor
 import org.oppia.android.data.backends.gae.NetworkSettings
-import org.oppia.android.testing.network.MockConceptCardService
+import org.oppia.android.data.backends.gae.api.SubtopicService
 import org.robolectric.annotation.LooperMode
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -16,11 +17,11 @@ import retrofit2.mock.MockRetrofit
 import retrofit2.mock.NetworkBehavior
 
 /**
- * Test for [ConceptCardService] retrofit instance using [MockConceptCardService]
+ * Test for [SubtopicService] retrofit instance using [MockSubtopicService]
  */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
-class ConceptCardServiceTest {
+class SubtopicServiceTest {
   private lateinit var mockRetrofit: MockRetrofit
   private lateinit var retrofit: Retrofit
 
@@ -42,20 +43,14 @@ class ConceptCardServiceTest {
   }
 
   @Test
-  fun testConceptCardService_usingFakeJson_deserializationSuccessful() {
-    val delegate = mockRetrofit.create(ConceptCardService::class.java)
-    val mockConceptCardService = MockConceptCardService(delegate)
+  fun testSubtopicService_usingFakeJson_deserializationSuccessful() {
+    val delegate = mockRetrofit.create(SubtopicService::class.java)
+    val mockSubtopicService = MockSubtopicService(delegate)
 
-    val skillIdList = ArrayList<String>()
-    skillIdList.add("1")
-    skillIdList.add("2")
-    skillIdList.add("3")
+    val subtopic = mockSubtopicService.getSubtopic("Subtopic 1", "randomId")
+    val subtopicResponse = subtopic.execute()
 
-    val skillIds = skillIdList.joinToString(separator = ", ")
-    val conceptCard = mockConceptCardService.getSkillContents(skillIds)
-    val conceptCardResponse = conceptCard.execute()
-
-    assertThat(conceptCardResponse.isSuccessful).isTrue()
-    assertThat(conceptCardResponse.body()!!.conceptCardDicts!!.size).isEqualTo(1)
+    assertThat(subtopicResponse.isSuccessful).isTrue()
+    assertThat(subtopicResponse.body()!!.subtopicTitle).isEqualTo("Subtopic 1")
   }
 }
