+++ data/src/test/java/org/oppia/android/data/gae/api/TopicServiceTest.kt
@@ -1,4 +1,4 @@
-package org.oppia.android.data.backends.gae.api
+package org.oppia.android.data.gae.api
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
@@ -6,9 +6,10 @@ import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.oppia.android.testing.network.MockTopicService
 import org.oppia.android.data.backends.gae.NetworkInterceptor
 import org.oppia.android.data.backends.gae.NetworkSettings
-import org.oppia.android.testing.network.MockQuestionPlayerService
+import org.oppia.android.data.backends.gae.api.TopicService
 import org.robolectric.annotation.LooperMode
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -16,11 +17,11 @@ import retrofit2.mock.MockRetrofit
 import retrofit2.mock.NetworkBehavior
 
 /**
- * Test for [QuestionPlayerService] retrofit instance using [MockQuestionPlayerService]
+ * Test for [TopicService] retrofit instance using [MockTopicService]
  */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
-class QuestionPlayerServiceTest {
+class TopicServiceTest {
   private lateinit var mockRetrofit: MockRetrofit
   private lateinit var retrofit: Retrofit
 
@@ -42,19 +43,14 @@ class QuestionPlayerServiceTest {
   }
 
   @Test
-  fun testQuestionPlayerService_usingFakeJson_deserializationSuccessful() {
-    val delegate = mockRetrofit.create(QuestionPlayerService::class.java)
-    val mockQuestionPlayerService = MockQuestionPlayerService(delegate)
+  fun testTopicService_usingFakeJson_deserializationSuccessful() {
+    val delegate = mockRetrofit.create(TopicService::class.java)
+    val mockTopicService = MockTopicService(delegate)
 
-    val skillIdList = ArrayList<String>()
-    skillIdList.add("1")
-    skillIdList.add("2")
-    skillIdList.add("3")
-    val skillIds = skillIdList.joinToString(separator = ", ")
-    val questionPlayer = mockQuestionPlayerService.getQuestionPlayerBySkillIds(skillIds, 10)
-    val questionPlayerResponse = questionPlayer.execute()
+    val topic = mockTopicService.getTopicByName("Topic1")
+    val topicResponse = topic.execute()
 
-    assertThat(questionPlayerResponse.isSuccessful).isTrue()
-    assertThat(questionPlayerResponse.body()!!.questions!!.size).isEqualTo(1)
+    assertThat(topicResponse.isSuccessful).isTrue()
+    assertThat(topicResponse.body()!!.topicName).isEqualTo("Topic1")
   }
 }
